+++ src/Uploadable/FileInfo/FileInfoInterface.php
@@ -28,7 +28,7 @@ interface FileInfoInterface
     public function getName();
 
     /**
-     * @return string|null
+     * @return int|null
      */
     public function getSize();